+++ Root/Particle.cxx
@@ -1,0 +1,32 @@
+#include "SusyNtuple/Particle.h"
+
+using namespace std;
+using Susy::Particle;
+
+/*--------------------------------------------------------------------------------*/
+// Copy constructor
+/*--------------------------------------------------------------------------------*/
+Particle::Particle(const Particle &rhs):
+  TLorentzVector(rhs),
+  idx(rhs.idx),
+  pt(rhs.pt),
+  eta(rhs.eta),
+  phi(rhs.phi),
+  m(rhs.m)
+{
+}
+/*--------------------------------------------------------------------------------*/
+// Assignment operator
+/*--------------------------------------------------------------------------------*/
+Particle& Particle::operator=(const Particle &rhs)
+{
+  if (this != &rhs) {
+    TLorentzVector::operator=(rhs);
+    idx  = rhs.idx; 
+    pt = rhs.pt; 
+    eta = rhs.eta; 
+    phi = rhs.phi; 
+    m = rhs.m; 
+  }
+  return *this;
+}
